+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- Created a quote button which goes to /b2b-quote/create flow
+- Added a app setting to enable/disable the create a quote button at the checkout page
+
 ## [0.7.0] - 2022-02-18
 
 ### Added